@@ -4,7 +4,7 @@
   "license": "MIT",
   "dependencies": {
     "isomorphic-fetch": "^2.2.1",
-    "next": "^1.1.1",
+    "next": "2.0.0-beta.5",
     "rebass": "^0.3.3",
     "reflexbox": "^2.2.3"
   },
