@@ -1,28 +1,30 @@
-import React from 'react'
-
-// Utilities
-import css from 'next/css'
+import React, { PropTypes } from 'react'
 
 function Circle ({number}) {
-  const style = css({
-    display: 'inline-block',
-    width: '30px',
-    height: '30px',
-    padding: '5px',
-    marginRight: '10px',
-    borderRadius: '50%',
-    backgroundColor: '#fff',
-    color: '#333',
-    textAlign: 'center'
-  })
-
-  return <span className={style}>{number}</span>
+  return (
+    <span>
+      {number}
+      <style jsx>{`
+        span {
+          display: inline-block;
+          width: 30px;
+          height: 30px;
+          padding: 5px;
+          margin-right: 10px;
+          border-radius: 50%;
+          background-color: #fff;
+          color: #333;
+          text-align: center;
+        }
+      `}</style>
+    </span>
+  )
 }
 
 Circle.propTypes = {
-  number: React.PropTypes.oneOfType([
-    React.PropTypes.number,
-    React.PropTypes.string
+  number: PropTypes.oneOfType([
+    PropTypes.number,
+    PropTypes.string
   ])
 }
 
