@@ -1,29 +1,28 @@
-import React from 'react'
-
-// Utilities
-import css from 'next/css'
+import React, { PropTypes } from 'react'
 
 // Components
 import ShowImage from './show-image'
 
 function ShowCard ({image}) {
-  const cardStyle = css({
-    width: '200px',
-    transition: '0.4s ease all',
-    ':hover': {
-      transform: 'scale(0.95)'
-    }
-  })
-
   return (
-    <div className={cardStyle}>
+    <div>
       <ShowImage src={image} preview fullWidth />
+      <style jsx>{`
+        div {
+          width: 200px;
+          transition: 0.4s ease all;
+        }
+
+        div:hover {
+          transform: scale(0.95);
+        }
+      `}</style>
     </div>
   )
 }
 
 ShowCard.propTypes = {
-  image: React.PropTypes.string.isRequired
+  image: PropTypes.string.isRequired
 }
 
 export default ShowCard
